fix: don't set $ExtraPipArgs to empty string
It seems that this gets passed to pip-wheels as a positional argument
and resulting in "ERROR: Invalid requirement: ''". Since we are not
checking return values from the pip command, this is silently ignored.
Not setting it at all(keeping the $null), in case where the $BuildEnv
is empty, seem to be working.

Fixes: eaf1cc85621c ("feat(build): Add initial support for custom build env..")

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>