fix: check sub-commands exit codes (#16)
* fix: don't set $ExtraPipArgs to empty string

It seems that this gets passed to pip-wheels as a positional argument
and resulting in "ERROR: Invalid requirement: ''". Since we are not
checking return values from the pip command, this is silently ignored.
Not setting it at all(keeping the $null), in case where the $BuildEnv
is empty, seem to be working.

Fixes: eaf1cc85621c ("feat(build): Add initial support for custom build env..")

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>

* fix: check return values from sub-commands

Currently the return value, e.g. from the pip-wheel command, is not checked
and the whole script can have $LASTEXITCODE set to 0, even though some
previous commands failed. As result to pipeline is not marked as failed.
Fix this by adding check for return values.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>

---------

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>